requirements.txt
torch>=2.0.0
torchvision>=0.15.0
dvc>=2.0.0
streamlit>=1.20.0
pandas
numpy
pyyaml
scikit-learn
matplotlib